[#4] Return additional error details when `:success?` is false.
The main methods of the library return a `:success?` key specifying if
the operation succeeded or not. But they don't give any details about
why the operation was unsuccessful. That information is only sent to
the logs.

This additional information is especially useful when a query violates
an integrity constratint (duplicate primary key, duplicate unique
column, NULL value for a non-NULL column, etc). Returning this
information to the caller allows it to handle the situation withou
easier, without needing additional queries to find out the issue.

Current implementation doesn't only distinguish between integrity
constraint violations, generic SQL errors and other kind of generic
errors (like missing parameters for the SQL query, etc). But this
could be extended in the future without changing the API.

Closes #4